fix(eslint-plugin): fix false positives for adjacent-overload-signatures regarding computed property names (#340)
Currently, the name of a method definition or a class property is always assumed to be a literal or an identifier. However, this fails to account for computed property names.
In those cases, we currently sometimes give the name for a method definition with a computed name as `undefined`, which as a result causes methods with differing computed property names to be marked as overloads of each other. In this PR, we slice the text of the computed property name out to handle these cases.